Fix timestamp in collected data, to be compliant with ISO8601
DISCLAIMER: this update can break existing log processing chains, like the
ELK stack.

The logged timestamp is expected to be ISO8601 compliant, to be easy to
parse with log processing chains.
However, before this fix, the logged timestamp represented the *local time*
and was not ISO8601 compliant.
To solve this problem, this fix aims at logging a *UTC* timestamp, which is
ISO8601 compliant.
In other words, the collected timestamp is now less human readable and more
processing chain friendly.